iteration for displaying numbers
this is some code that has been written in order for the display to run and show numbers. It isn't working for me, but I somewhat understand how it works. I will take sometime to see how I can get it to work with the code that I've already written